+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- `true` and `false` can no longer be used as pattern matching variables.
+  A syntax error will hint about using Gleam's `True` and `False` values instead.
 - You can now remove build artifacts using the new `gleam clean` command.
 - The `compile-package` can now generate `package.app` files and compile source
   modules to `.beam` bytecode files.